Fixed typo in ValidationExtensions class name
